Decoupling for Unit tests
