+++ grails-app/jobs/de/iteratec/osm/report/external/GraphiteReportJob.groovy
@@ -1,25 +1,25 @@
 package de.iteratec.osm.report.external
 
-import de.iteratec.osm.measurement.environment.wptserver.EventResultPersisterService
-import de.iteratec.osm.measurement.environment.wptserver.WptResultXml
+
+import de.iteratec.osm.result.EventResult
 import de.iteratec.osm.util.PerformanceLoggingService
 import org.quartz.JobExecutionContext
-import static de.iteratec.osm.util.PerformanceLoggingService.LogLevel.DEBUG
 
 class GraphiteReportJob {
 
-    EventResultPersisterService eventResultPersisterService
-    PerformanceLoggingService performanceLoggingService
+    MetricReportingService metricReportingService
 
     static triggers = {}
 
     def execute(JobExecutionContext context) {
-        WptResultXml resultXml = context.mergedJobDataMap.get('resultXml')
-        Long jobId = context.mergedJobDataMap.getLong('jobId')
-        if(resultXml && jobId) {
-            performanceLoggingService.logExecutionTime(DEBUG, "Inform Dependents for Job $jobId", 1) {
-                eventResultPersisterService.informDependents(resultXml, jobId)
-            }
+        EventResult result = context.mergedJobDataMap.get('eventResult')
+
+        try {
+            metricReportingService.reportEventResultToGraphite(result)
+        } catch (GraphiteComunicationFailureException gcfe) {
+            log.error("Can't report EventResult to graphite-server: ${gcfe.message}")
+        } catch (Exception e) {
+            log.error("An error occurred while reporting EventResult to graphite.", e)
         }
     }
 }
