+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/PersistConnectivityInNewEventResultSpec.groovy
@@ -28,6 +28,7 @@ import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.measurement.schedule.JobDaoService
 import de.iteratec.osm.measurement.schedule.JobGroup
 import de.iteratec.osm.measurement.script.Script
+import de.iteratec.osm.report.external.GraphiteReportService
 import de.iteratec.osm.report.external.MetricReportingService
 import de.iteratec.osm.result.*
 import de.iteratec.osm.util.PerformanceLoggingService
@@ -78,6 +79,8 @@ class PersistConnectivityInNewEventResultSpec extends Specification implements B
             getMaxValidLoadtime() >> DEFAULT_MAX_VALID_LOADTIME
             getMinValidLoadtime() >> DEFAULT_MIN_VALID_LOADTIME
         }
+
+        service.graphiteReportService = Mock(GraphiteReportService)
     }
 
     void setupSpec() {
