+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/PersistScreenshotDependentWptMetricsSpec.groovy
@@ -28,6 +28,7 @@ import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.measurement.schedule.JobDaoService
 import de.iteratec.osm.measurement.schedule.JobGroup
 import de.iteratec.osm.measurement.script.Script
+import de.iteratec.osm.report.external.GraphiteReportService
 import de.iteratec.osm.report.external.MetricReportingService
 import de.iteratec.osm.result.*
 import de.iteratec.osm.util.PerformanceLoggingService
@@ -247,6 +248,8 @@ class PersistScreenshotDependentWptMetricsSpec extends Specification implements 
             getMaxValidLoadtime() >> DEFAULT_MAX_VALID_LOADTIME
             getMinValidLoadtime() >> DEFAULT_MIN_VALID_LOADTIME
         }
+
+        service.graphiteReportService = Mock(GraphiteReportService)
     }
 
 }
