+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/PersistingNewEventResultsSpec.groovy
@@ -25,6 +25,7 @@ import de.iteratec.osm.measurement.environment.WebPageTestServer
 import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.measurement.schedule.JobDaoService
 import de.iteratec.osm.measurement.schedule.JobGroup
+import de.iteratec.osm.report.external.GraphiteReportService
 import de.iteratec.osm.report.external.MetricReportingService
 import de.iteratec.osm.result.*
 import de.iteratec.osm.util.PerformanceLoggingService
@@ -66,6 +67,8 @@ class PersistingNewEventResultsSpec extends Specification implements BuildDataTe
             getMaxValidLoadtime() >> DEFAULT_MAX_VALID_LOADTIME
             getMinValidLoadtime() >> DEFAULT_MIN_VALID_LOADTIME
         }
+
+        service.graphiteReportService = Mock(GraphiteReportService)
     }
 
     void "result persistance with old (single step) WPT server"(String fileName, String jobLabel, String pageName) {
