+++ src/test/groovy/de/iteratec/osm/measurement/environment/wptserver/PersistingNewEventResultsWithNoMedianOptionTestSpec.groovy
@@ -28,6 +28,7 @@ import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.measurement.schedule.JobDaoService
 import de.iteratec.osm.measurement.schedule.JobGroup
 import de.iteratec.osm.measurement.script.Script
+import de.iteratec.osm.report.external.GraphiteReportService
 import de.iteratec.osm.result.*
 import de.iteratec.osm.util.PerformanceLoggingService
 import grails.buildtestdata.BuildDataTest
@@ -66,6 +67,7 @@ class PersistingNewEventResultsWithNoMedianOptionTestSpec extends Specification 
             getMaxValidLoadtime() >> DEFAULT_MAX_VALID_LOADTIME
             getMinValidLoadtime() >> DEFAULT_MIN_VALID_LOADTIME
         }
+        service.graphiteReportService = Mock(GraphiteReportService)
 
         WebPageTestServer wptServer = WebPageTestServer.build(baseUrl: "http://wpt.org")
         Location.build(uniqueIdentifierForServer: locationIdentifier, wptServer: wptServer)
